+++ src/main/java/com/jiebao/platfrom/check/service/IGradeService.java
@@ -17,7 +17,7 @@ import java.util.List;
  * @since 2020-07-28
  */
 public interface IGradeService extends IService<Grade> {
-    JiebaoResponse addGrade(String menusId, Integer number, String yearDate, String deptId);    //扣分绑定
+    JiebaoResponse addGrade(String menusId, Double number, String yearDate, String deptId);    //扣分绑定
 
     JiebaoResponse commit(String yearDate, String deptId);//最后提交  分数统计生成表
